partially resolve #402: move analytical.py into tests
